requirements.txt
Django==1.10
django-import-export==0.5.1
djangorestframework==3.5.3
Markdown==2.6.3
psycopg2==2.7.1
Pygments==2.0.2
